requirements.txt
dvc
numpy 
pyyaml
dill
evidently
pandas
seaborn
pymongo
from-root
matplotlib
python-dotenv
xgboost
scikit-learn
-e .